Adjusted ReadSPI to return Ints instead of Bytes, this should adjust for shifting required to complete concatenation of data.

+++ src/main/scala/ADS1256.scala
@@ -240,12 +240,12 @@ object ADS1256 {
     }
   }
 
-  def ReadSPI(n: Int): List[Byte] = {
-    var result = ListBuffer[Byte]()
+  def ReadSPI(n: Int): List[Int] = {
+    var result = ListBuffer[Int]()
 
     for (i <- Range(0,n)) {
       // Write always returns an Array of Byte, even if it's just one byte. Pull the byte out and add it to the List
-      result += spidev.write(0x00.toByte)(0)
+      result += spidev.write(0x00.toByte)(0).toInt
     }
     result.toList
   }
